[Bug 666329] Empty database after a little while
Use gnc_dbi_safe_sync_all (which doesn't lose data if the save partion
errors out) for all be->sync.

This is a bit of a half-way solution; we really shouldn't ever have to do
this except when creating a new database, but fixing that will take major
surgery.

BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@21831 57a11ea4-9604-0410-9ed3-97b8803252fd